requirements.txt
numpy
opencv-contrib-python
PyYAML
requests
tinytuya
twilio
flask
